requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pgvector>=0.3.0
openai>=1.30.0
pydantic-settings>=2.2.1
stripe>=9.0.0
httpx>=0.24.0

python-dateutil>=2.8
sentry-sdk[fastapi]>=2.0.0
python-json-logger>=2.0.0
ddgs>=1.0.0
resend>=0.8.0
